Move checkExpiredFioDomains into FioServices
